+++ src/main/java/md/ceiti/spring/entity/dto/country/CountryDto.java
@@ -1,15 +1,15 @@
 package md.ceiti.spring.entity.dto.country;
 
 public class CountryDto {
-    private Long countryId;
+    private Integer countryId;
     private String countryName;
 
-    public CountryDto(Long countryId, String countryName) {
+    public CountryDto(Integer countryId, String countryName) {
         this.countryId = countryId;
         this.countryName = countryName;
     }
 
-    public Long getCountryId() {
+    public Integer getCountryId() {
         return countryId;
     }
